Move static methods to a separate class
This seems to confuse TestNG, which tries to invoke some of the public
static methods during setup
