+++ presto-cassandra/src/test/java/com/facebook/presto/cassandra/TestCassandraConnector.java
@@ -32,37 +32,17 @@ import com.facebook.presto.spi.SchemaTableName;
 import com.facebook.presto.spi.SchemaTablePrefix;
 import com.facebook.presto.spi.TupleDomain;
 import com.facebook.presto.spi.type.Type;
-import com.google.common.base.Charsets;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
-import com.google.common.primitives.Ints;
-import me.prettyprint.cassandra.model.BasicColumnDefinition;
-import me.prettyprint.cassandra.serializers.BytesArraySerializer;
-import me.prettyprint.cassandra.serializers.IntegerSerializer;
-import me.prettyprint.cassandra.serializers.LongSerializer;
-import me.prettyprint.cassandra.serializers.StringSerializer;
-import me.prettyprint.cassandra.serializers.UUIDSerializer;
-import me.prettyprint.hector.api.Cluster;
-import me.prettyprint.hector.api.Keyspace;
-import me.prettyprint.hector.api.ddl.ColumnDefinition;
-import me.prettyprint.hector.api.ddl.ColumnFamilyDefinition;
-import me.prettyprint.hector.api.ddl.ColumnType;
-import me.prettyprint.hector.api.ddl.ComparatorType;
-import me.prettyprint.hector.api.ddl.KeyspaceDefinition;
-import me.prettyprint.hector.api.factory.HFactory;
-import me.prettyprint.hector.api.mutation.Mutator;
-import org.cassandraunit.model.StrategyModel;
 import org.cassandraunit.utils.EmbeddedCassandraServerHelper;
 import org.testng.annotations.AfterMethod;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
 
-import java.nio.ByteBuffer;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
-import java.util.UUID;
 
+import static com.facebook.presto.cassandra.CassandraTestingUtils.createTestData;
 import static com.facebook.presto.cassandra.util.Types.checkType;
 import static com.facebook.presto.spi.type.BigintType.BIGINT;
 import static com.facebook.presto.spi.type.BooleanType.BOOLEAN;
@@ -91,8 +71,6 @@ public class TestCassandraConnector
     protected SchemaTableName table;
     protected SchemaTableName tableUnpartitioned;
     protected SchemaTableName invalidTable;
-    private static final String CLUSTER_NAME = "TestCluster";
-    private static final String HOST = "localhost:9160";
 
     @BeforeClass
     public void setup()
@@ -230,11 +208,6 @@ public class TestCassandraConnector
         assertEquals(rowNumber, 9);
     }
 
-    private String toUtf8String(byte[] keys)
-    {
-        return new String(keys, Charsets.UTF_8);
-    }
-
     private static void assertReadFields(RecordCursor cursor, List<ColumnMetadata> schema)
     {
         for (int columnIndex = 0; columnIndex < schema.size(); columnIndex++) {
@@ -293,144 +266,5 @@ public class TestCassandraConnector
             i++;
         }
         return index.build();
-    }
-
-    public static Keyspace createOrReplaceKeyspace(String keyspaceName)
-    {
-        return createOrReplaceKeyspace(keyspaceName, ImmutableList.<ColumnFamilyDefinition>of());
-    }
-
-    public static Keyspace createOrReplaceKeyspace(String keyspaceName, List<ColumnFamilyDefinition> columnFamilyDefinitions)
-    {
-        Cluster cluster = getOrCreateCluster();
-
-        KeyspaceDefinition keyspaceDefinition = HFactory.createKeyspaceDefinition(
-                keyspaceName,
-                StrategyModel.SIMPLE_STRATEGY.value(),
-                1,
-                columnFamilyDefinitions);
-
-        if (cluster.describeKeyspace(keyspaceName) != null) {
-            cluster.dropKeyspace(keyspaceName, true);
-        }
-        cluster.addKeyspace(keyspaceDefinition, true);
-        return HFactory.createKeyspace(keyspaceName, cluster);
-    }
-
-    public static void createTestData(String keyspaceName, String columnFamilyName)
-    {
-        List<ColumnFamilyDefinition> columnFamilyDefinitions = createColumnFamilyDefinitions(keyspaceName, columnFamilyName);
-        Keyspace keyspace = createOrReplaceKeyspace(keyspaceName, columnFamilyDefinitions);
-
-        Mutator<String> mutator = HFactory.createMutator(keyspace, StringSerializer.get());
-
-        long timestamp = System.currentTimeMillis();
-        for (int rowNumber = 1; rowNumber < 10; rowNumber++) {
-            addRow(columnFamilyName, mutator, timestamp, rowNumber);
-        }
-        mutator.execute();
-    }
-
-    private static Cluster getOrCreateCluster()
-    {
-        return HFactory.getOrCreateCluster(CLUSTER_NAME, HOST);
-    }
-
-    private static void addRow(String columnFamilyName, Mutator<String> mutator, long timestamp, int rowNumber)
-    {
-        String key = String.format("key %04d", rowNumber);
-        mutator.addInsertion(
-                key,
-                columnFamilyName,
-                HFactory.createColumn(
-                        "t_utf8",
-                        "utf8 " + rowNumber,
-                        timestamp,
-                        StringSerializer.get(),
-                        StringSerializer.get()));
-        mutator.addInsertion(
-                key,
-                columnFamilyName,
-                HFactory.createColumn(
-                        "t_bytes",
-                        Ints.toByteArray(rowNumber),
-                        timestamp,
-                        StringSerializer.get(),
-                        BytesArraySerializer.get()));
-        mutator.addInsertion(
-                key,
-                columnFamilyName,
-                HFactory.createColumn(
-                        "t_integer",
-                        rowNumber,
-                        timestamp,
-                        StringSerializer.get(),
-                        IntegerSerializer.get()));
-        mutator.addInsertion(
-                key,
-                columnFamilyName,
-                HFactory.createColumn(
-                        "t_long",
-                        1000L + rowNumber,
-                        timestamp,
-                        StringSerializer.get(),
-                        LongSerializer.get()));
-        mutator.addInsertion(
-                key,
-                columnFamilyName,
-                HFactory.createColumn(
-                        "t_uuid",
-                        UUID.fromString(String.format("00000000-0000-0000-0000-%012d", rowNumber)),
-                        timestamp,
-                        StringSerializer.get(),
-                        UUIDSerializer.get()));
-        mutator.addInsertion(
-                key,
-                columnFamilyName,
-                HFactory.createColumn(
-                        "t_lexical_uuid",
-                        UUID.fromString(String.format("00000000-0000-0000-0000-%012d", rowNumber)),
-                        timestamp,
-                        StringSerializer.get(),
-                        UUIDSerializer.get()));
-    }
-
-    private static List<ColumnFamilyDefinition> createColumnFamilyDefinitions(String keyspaceName, String columnFamilyName)
-    {
-        List<ColumnFamilyDefinition> columnFamilyDefinitions = new ArrayList<>();
-
-        ImmutableList.Builder<ColumnDefinition> columnsDefinition = ImmutableList.builder();
-
-        columnsDefinition.add(createColumnDefinition("t_utf8", ComparatorType.UTF8TYPE));
-        columnsDefinition.add(createColumnDefinition("t_bytes", ComparatorType.BYTESTYPE));
-        columnsDefinition.add(createColumnDefinition("t_integer", ComparatorType.INTEGERTYPE));
-        columnsDefinition.add(createColumnDefinition("t_int32", ComparatorType.INT32TYPE));
-        columnsDefinition.add(createColumnDefinition("t_long", ComparatorType.LONGTYPE));
-        columnsDefinition.add(createColumnDefinition("t_boolean", ComparatorType.BOOLEANTYPE));
-        columnsDefinition.add(createColumnDefinition("t_uuid", ComparatorType.UUIDTYPE));
-        columnsDefinition.add(createColumnDefinition("t_lexical_uuid", ComparatorType.LEXICALUUIDTYPE));
-
-        ColumnFamilyDefinition cfDef = HFactory.createColumnFamilyDefinition(
-                keyspaceName,
-                columnFamilyName,
-                ComparatorType.UTF8TYPE,
-                columnsDefinition.build());
-
-        cfDef.setColumnType(ColumnType.STANDARD);
-        cfDef.setComment("presto test table");
-
-        cfDef.setKeyValidationClass(ComparatorType.UTF8TYPE.getTypeName());
-
-        columnFamilyDefinitions.add(cfDef);
-
-        return columnFamilyDefinitions;
-    }
-
-    private static BasicColumnDefinition createColumnDefinition(String columnName, ComparatorType type)
-    {
-        BasicColumnDefinition columnDefinition = new BasicColumnDefinition();
-        columnDefinition.setName(ByteBuffer.wrap(columnName.getBytes(Charsets.UTF_8)));
-        columnDefinition.setValidationClass(type.getClassName());
-        return columnDefinition;
     }
 }
