+++ presto-cassandra/src/test/java/com/facebook/presto/cassandra/TestCassandraIntegrationSmokeTest.java
@@ -19,6 +19,8 @@ import org.testng.annotations.Test;
 import static com.facebook.presto.cassandra.CassandraQueryRunner.createCassandraQueryRunner;
 import static com.facebook.presto.cassandra.CassandraQueryRunner.createCassandraSession;
 import static com.facebook.presto.cassandra.CassandraQueryRunner.createSampledSession;
+import static com.facebook.presto.cassandra.CassandraTestingUtils.createOrReplaceKeyspace;
+import static com.facebook.presto.cassandra.CassandraTestingUtils.createTestData;
 import static io.airlift.tpch.TpchTable.ORDERS;
 
 @Test(singleThreaded = true)
@@ -34,9 +36,8 @@ public class TestCassandraIntegrationSmokeTest
     @Test
     public void testStringPartitionKey()
     {
-        TestCassandraConnector.createOrReplaceKeyspace("example");
-        TestCassandraConnector.createTestData("example", "test_key_text");
+        createOrReplaceKeyspace("example");
+        createTestData("example", "test_key_text");
         queryRunner.execute(createCassandraSession("example"), "select * from example.test_key_text where key='key 0001'");
-
     }
 }
